Improve code

+++ weather/include/weather/weather_data.hh
@@ -11,7 +11,7 @@ struct weather_data {
     unsigned int deg = 0;  // wind direction in degrees
   } wind;
   struct {
-    float all = 0;   // cloud coverage in percent
+    unsigned all = 0;   // cloud coverage in percent
   } clouds;
 
 public:
@@ -26,7 +26,7 @@ public:
   float get_wind_speed_kph() const {
     return wind.speed;
   }
-  float get_cloud_coverage_percent() const {
+  unsigned get_cloud_coverage_percent() const {
     return clouds.all;
   }
 };
